requirements.txt
xgboost==1.0.2
numpy==1.18.2
pandas==0.25.3
jieba==0.42.1
astropy==4.0.1
scipy==1.4.1
lxml==4.5.0
cssselect==1.1.0
QtAwesome==0.7.0
requests==2.22.0
snownlp==0.12.3
phik==0.9.9
matplotlib==3.2.0
confuse==1.0.0
wordcloud==1.6.0
Pillow==7.1.1
PyQt5==5.14.2
python_dateutil==2.8.1
scikit_learn==0.22.2.post1
sip==5.1.2
